adding vector database clas
